feat : Logger.cs -> Would log the messages to a file, Constants.cs -> would have all the constants, User Authentication implemented.
ExpenseTrackerUnitTests -> Tests written

+++ CSharp_Assignments_Phase2/ExpenseTrackerConsoleApplication/Expense.cs
@@ -20,7 +20,7 @@ namespace ExpenseTrackerConsoleApplication
         /// <param name="amountModes">AmountModes</param>
         /// <param name="note">Notes</param>
         /// <param name="account">Account</param>
-        public Expense(DateOnly date, string category,int account, ModesOfCash amountModes, string note, double amount = 0)
+        public Expense(DateOnly date, string category,int account, ModesOfCash amountModes, string note, decimal amount = 0)
         {
             DateOnly = date;
             Category = category;
